requirements-dev.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
gunicorn==21.2.0
# psycopg2-binary==2.9.7  # Commented out for development - use SQLite instead
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
